(unstable) [fixed] Items with large oid (i.e. PC mission terms) are now sliceable  - Patch by Cebot
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/branches/unstable@6147 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/tangible/tool/smuggler/PrecisionLaserKnifeImplementation.cpp
@@ -23,7 +23,7 @@ int PrecisionLaserKnifeImplementation::handleObjectMenuSelect(CreatureObject* pl
 	if (selectedID != 20)
 		return TangibleObjectImplementation::handleObjectMenuSelect(player, selectedID);
 
-	uint32 targetID = player->getTargetID();
+	uint64 targetID = player->getTargetID();
 	ZoneServer* zs = player->getZoneServer();
 	ManagedReference<TangibleObject*> target = dynamic_cast<TangibleObject*>(zs->getObject(targetID, true));
